Add ControllerInterfaceBase class and use it in Controller Manager.
