+++ controller_interface/include/controller_interface/controller_interface.hpp
@@ -19,128 +19,51 @@
 #include <string>
 #include <vector>
 
+#include "controller_interface/controller_interface_base.hpp"
 #include "controller_interface/visibility_control.h"
-
-#include "hardware_interface/loaned_command_interface.hpp"
-#include "hardware_interface/loaned_state_interface.hpp"
-
-#include "rclcpp/rclcpp.hpp"
-#include "rclcpp_lifecycle/lifecycle_node.hpp"
+#include "hardware_interface/handle.hpp"
 
 namespace controller_interface
 {
-using CallbackReturn = rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn;
-
-enum class return_type : std::uint8_t
-{
-  OK = 0,
-  ERROR = 1,
-};
-
-/// Indicating which interfaces are to be claimed.
-/**
- * One might either claim all available command/state interfaces,
- * specifying a set of individual interfaces,
- * or none at all.
- */
-enum class interface_configuration_type : std::uint8_t
-{
-  ALL = 0,
-  INDIVIDUAL = 1,
-  NONE = 2,
-};
-
-/// Configuring what command/state interfaces to claim.
-struct InterfaceConfiguration
-{
-  interface_configuration_type type;
-  std::vector<std::string> names = {};
-};
-
-class ControllerInterface : public rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface
+class ControllerInterface : public controller_interface::ControllerInterfaceBase
 {
 public:
   CONTROLLER_INTERFACE_PUBLIC
-  ControllerInterface() = default;
+  ControllerInterface();
 
   CONTROLLER_INTERFACE_PUBLIC
   virtual ~ControllerInterface() = default;
 
-  CONTROLLER_INTERFACE_PUBLIC
-  virtual InterfaceConfiguration command_interface_configuration() const = 0;
-
-  CONTROLLER_INTERFACE_PUBLIC
-  virtual InterfaceConfiguration state_interface_configuration() const = 0;
-
-  CONTROLLER_INTERFACE_PUBLIC
-  void assign_interfaces(
-    std::vector<hardware_interface::LoanedCommandInterface> && command_interfaces,
-    std::vector<hardware_interface::LoanedStateInterface> && state_interfaces);
-
-  CONTROLLER_INTERFACE_PUBLIC
-  void release_interfaces();
-
-  CONTROLLER_INTERFACE_PUBLIC
-  virtual return_type init(
-    const std::string & controller_name, const std::string & namespace_ = "",
-    const rclcpp::NodeOptions & node_options =
-      rclcpp::NodeOptions()
-        .allow_undeclared_parameters(true)
-        .automatically_declare_parameters_from_overrides(true));
-
-  /// Custom configure method to read additional parameters for controller-nodes
-  /*
-   * Override default implementation for configure of LifecycleNode to get parameters.
-   */
-  CONTROLLER_INTERFACE_PUBLIC
-  const rclcpp_lifecycle::State & configure();
-
-  /// Extending interface with initialization method which is individual for each controller
-  CONTROLLER_INTERFACE_PUBLIC
-  virtual CallbackReturn on_init() = 0;
-
-  CONTROLLER_INTERFACE_PUBLIC
-  virtual return_type update(const rclcpp::Time & time, const rclcpp::Duration & period) = 0;
-
-  CONTROLLER_INTERFACE_PUBLIC
-  std::shared_ptr<rclcpp_lifecycle::LifecycleNode> get_node();
-
-  CONTROLLER_INTERFACE_PUBLIC
-  std::shared_ptr<rclcpp_lifecycle::LifecycleNode> get_node() const;
-
-  CONTROLLER_INTERFACE_PUBLIC
-  const rclcpp_lifecycle::State & get_state() const;
-
-  CONTROLLER_INTERFACE_PUBLIC
-  unsigned int get_update_rate() const;
-
-  /// Declare and initialize a parameter with a type.
   /**
+   * Controller is not chainable.
    *
-   * Wrapper function for templated node's declare_parameter() which checks if
-   * parameter is already declared.
-   * For use in all components that inherit from ControllerInterface
+   * \returns false.
    */
-  template <typename ParameterT>
-  auto auto_declare(const std::string & name, const ParameterT & default_value)
-  {
-    if (!node_->has_parameter(name))
-    {
-      return node_->declare_parameter<ParameterT>(name, default_value);
-    }
-    else
-    {
-      return node_->get_parameter(name).get_value<ParameterT>();
-    }
-  }
+  CONTROLLER_INTERFACE_PUBLIC
+  bool is_chainable() const final;
 
-protected:
-  std::vector<hardware_interface::LoanedCommandInterface> command_interfaces_;
-  std::vector<hardware_interface::LoanedStateInterface> state_interfaces_;
-  unsigned int update_rate_ = 0;
+  /**
+   * Controller has not reference interfaces.
+   *
+   * \returns empty list.
+   */
+  CONTROLLER_INTERFACE_PUBLIC
+  std::vector<hardware_interface::CommandInterface> export_reference_interfaces() final;
 
-private:
-  std::shared_ptr<rclcpp_lifecycle::LifecycleNode> node_;
+  /**
+   * Controller is not chainable, therefore no chained mode can be set.
+   *
+   * \returns false;
+   */
+  CONTROLLER_INTERFACE_PUBLIC
+  bool set_chained_mode(bool chained_mode) final;
+  /**
+   * Controller can not be in chained mode.
+   *
+   * \returns false.
+   */
+  CONTROLLER_INTERFACE_PUBLIC
+  bool is_in_chained_mode() const final;
 };
 
 using ControllerInterfaceSharedPtr = std::shared_ptr<ControllerInterface>;
