+++ controller_manager/src/controller_manager.cpp
@@ -30,8 +30,9 @@
 namespace
 {  // utility
 
-static constexpr const char * kControllerInterfaceName = "controller_interface";
-static constexpr const char * kControllerInterface = "controller_interface::ControllerInterface";
+static constexpr const char * kControllerInterfaceBaseName = "controller_interface";
+static constexpr const char * kControllerInterfaceBase =
+  "controller_interface::ControllerInterfaceBase";
 
 // Changed services history QoS to keep all so we don't lose any client service calls
 static const rmw_qos_profile_t rmw_qos_profile_services_hist_keep_all = {
@@ -45,24 +46,24 @@ static const rmw_qos_profile_t rmw_qos_profile_services_hist_keep_all = {
   RMW_QOS_LIVELINESS_LEASE_DURATION_DEFAULT,
   false};
 
-inline bool is_controller_inactive(const controller_interface::ControllerInterface & controller)
+inline bool is_controller_inactive(const controller_interface::ControllerInterfaceBase & controller)
 {
   return controller.get_state().id() == lifecycle_msgs::msg::State::PRIMARY_STATE_INACTIVE;
 }
 
 inline bool is_controller_inactive(
-  const controller_interface::ControllerInterfaceSharedPtr & controller)
+  const controller_interface::ControllerInterfaceBaseSharedPtr & controller)
 {
   return is_controller_inactive(*controller);
 }
 
-inline bool is_controller_active(const controller_interface::ControllerInterface & controller)
+inline bool is_controller_active(const controller_interface::ControllerInterfaceBase & controller)
 {
   return controller.get_state().id() == lifecycle_msgs::msg::State::PRIMARY_STATE_ACTIVE;
 }
 
 inline bool is_controller_active(
-  const controller_interface::ControllerInterfaceSharedPtr & controller)
+  const controller_interface::ControllerInterfaceBaseSharedPtr & controller)
 {
   return is_controller_active(*controller);
 }
@@ -91,8 +92,8 @@ ControllerManager::ControllerManager(
 : rclcpp::Node(manager_node_name, namespace_, get_cm_node_options()),
   resource_manager_(std::make_unique<hardware_interface::ResourceManager>()),
   executor_(executor),
-  loader_(std::make_shared<pluginlib::ClassLoader<controller_interface::ControllerInterface>>(
-    kControllerInterfaceName, kControllerInterface))
+  loader_(std::make_shared<pluginlib::ClassLoader<controller_interface::ControllerInterfaceBase>>(
+    kControllerInterfaceBaseName, kControllerInterfaceBase))
 {
   if (!get_parameter("update_rate", update_rate_))
   {
@@ -118,8 +119,8 @@ ControllerManager::ControllerManager(
 : rclcpp::Node(manager_node_name, namespace_, get_cm_node_options()),
   resource_manager_(std::move(resource_manager)),
   executor_(executor),
-  loader_(std::make_shared<pluginlib::ClassLoader<controller_interface::ControllerInterface>>(
-    kControllerInterfaceName, kControllerInterface))
+  loader_(std::make_shared<pluginlib::ClassLoader<controller_interface::ControllerInterfaceBase>>(
+    kControllerInterfaceBaseName, kControllerInterfaceBase))
 {
   init_services();
 }
@@ -226,7 +227,7 @@ void ControllerManager::init_services()
       rmw_qos_profile_services_hist_keep_all, best_effort_callback_group_);
 }
 
-controller_interface::ControllerInterfaceSharedPtr ControllerManager::load_controller(
+controller_interface::ControllerInterfaceBaseSharedPtr ControllerManager::load_controller(
   const std::string & controller_name, const std::string & controller_type)
 {
   RCLCPP_INFO(get_logger(), "Loading controller '%s'", controller_name.c_str());
@@ -251,7 +252,7 @@ controller_interface::ControllerInterfaceSharedPtr ControllerManager::load_contr
   return add_controller_impl(controller_spec);
 }
 
-controller_interface::ControllerInterfaceSharedPtr ControllerManager::load_controller(
+controller_interface::ControllerInterfaceBaseSharedPtr ControllerManager::load_controller(
   const std::string & controller_name)
 {
   const std::string param_name = controller_name + ".type";
@@ -680,7 +681,7 @@ controller_interface::return_type ControllerManager::switch_controller(
   return controller_interface::return_type::OK;
 }
 
-controller_interface::ControllerInterfaceSharedPtr ControllerManager::add_controller_impl(
+controller_interface::ControllerInterfaceBaseSharedPtr ControllerManager::add_controller_impl(
   const ControllerSpec & controller)
 {
   // lock controllers
@@ -980,7 +981,7 @@ void ControllerManager::list_controller_types_srv_cb(
   for (const auto & cur_type : cur_types)
   {
     response->types.push_back(cur_type);
-    response->base_classes.push_back(kControllerInterface);
+    response->base_classes.push_back(kControllerInterfaceBase);
     RCLCPP_DEBUG(get_logger(), "%s", cur_type.c_str());
   }
 
@@ -1172,11 +1173,11 @@ void ControllerManager::reload_controller_libraries_service_cb(
   assert(loaded_controllers.empty());
 
   // Force a reload on all the PluginLoaders (internally, this recreates the plugin loaders)
-  loader_ = std::make_shared<pluginlib::ClassLoader<controller_interface::ControllerInterface>>(
-    kControllerInterfaceName, kControllerInterface);
+  loader_ = std::make_shared<pluginlib::ClassLoader<controller_interface::ControllerInterfaceBase>>(
+    kControllerInterfaceBaseName, kControllerInterfaceBase);
   RCLCPP_INFO(
     get_logger(), "Controller manager: reloaded controller libraries for '%s'",
-    kControllerInterfaceName);
+    kControllerInterfaceBaseName);
 
   response->ok = true;
