+++ controller_manager/test/test_controller/test_controller.cpp
@@ -101,4 +101,5 @@ void TestController::set_state_interface_configuration(
 
 #include "pluginlib/class_list_macros.hpp"
 
-PLUGINLIB_EXPORT_CLASS(test_controller::TestController, controller_interface::ControllerInterface)
+PLUGINLIB_EXPORT_CLASS(
+  test_controller::TestController, controller_interface::ControllerInterfaceBase)
