+++ controller_manager/test/test_controller/test_controller.xml
@@ -1,6 +1,6 @@
 <library path="test_controller">
 
-  <class name="controller_manager/test_controller" type="test_controller::TestController" base_class_type="controller_interface::ControllerInterface">
+  <class name="controller_manager/test_controller" type="test_controller::TestController" base_class_type="controller_interface::ControllerInterfaceBase">
     <description>
       Controller used for testing
     </description>
