+++ controller_manager/test/test_controller_failed_init/test_controller_failed_init.cpp
@@ -49,4 +49,5 @@ controller_interface::return_type TestControllerFailedInit::update(
 #include "pluginlib/class_list_macros.hpp"
 
 PLUGINLIB_EXPORT_CLASS(
-  test_controller_failed_init::TestControllerFailedInit, controller_interface::ControllerInterface)
+  test_controller_failed_init::TestControllerFailedInit,
+  controller_interface::ControllerInterfaceBase)
