+++ controller_manager/test/test_controller_failed_init/test_controller_failed_init.xml
@@ -1,6 +1,6 @@
 <library path="test_controller_failed_init">
 
-  <class name="controller_manager/test_controller_failed_init" type="test_controller_failed_init::TestControllerFailedInit" base_class_type="controller_interface::ControllerInterface">
+  <class name="controller_manager/test_controller_failed_init" type="test_controller_failed_init::TestControllerFailedInit" base_class_type="controller_interface::ControllerInterfaceBase">
     <description>
       Controller used for testing
     </description>
