+++ controller_manager/test/test_controller_manager_srvs.cpp
@@ -48,7 +48,7 @@ TEST_F(TestControllerManagerSrvs, list_controller_types)
   ASSERT_EQ(controller_types, result->base_classes.size());
   ASSERT_THAT(result->types, ::testing::Contains(test_controller::TEST_CONTROLLER_CLASS_NAME));
   ASSERT_THAT(
-    result->base_classes, ::testing::Contains("controller_interface::ControllerInterface"));
+    result->base_classes, ::testing::Contains("controller_interface::ControllerInterfaceBase"));
 }
 
 TEST_F(TestControllerManagerSrvs, list_controllers_srv)
