+++ controller_manager/test/test_controller_with_interfaces/test_controller_with_interfaces.xml
@@ -1,6 +1,6 @@
 <library path="test_controller_with_interfaces">
 
-  <class name="controller_manager/test_controller_with_interfaces" type="test_controller_with_interfaces::TestControllerWithInterfaces" base_class_type="controller_interface::ControllerInterface">
+  <class name="controller_manager/test_controller_with_interfaces" type="test_controller_with_interfaces::TestControllerWithInterfaces" base_class_type="controller_interface::ControllerInterfaceBase">
     <description>
       Controller used for testing
     </description>
